+++ pkg/querier/queryrange/limits_test.go
@@ -31,14 +31,6 @@ func TestLimits(t *testing.T) {
 	require.Equal(t, l.QuerySplitDuration("b"), time.Duration(0))
 
 	wrapped := WithDefaultLimits(l, queryrange.Config{
-		SplitQueriesByDay: true,
-	})
-
-	require.Equal(t, wrapped.QuerySplitDuration("a"), time.Minute)
-	require.Equal(t, wrapped.QuerySplitDuration("b"), 24*time.Hour)
-
-	wrapped = WithDefaultLimits(l, queryrange.Config{
-		SplitQueriesByDay:      true, // should be overridden by SplitQueriesByInterval
 		SplitQueriesByInterval: time.Hour,
 	})